Compass-like arrow
- PointToHotel script deleted
- ArrowToWorkbench script added
