+++ Assets/Resources/Prefabs/UI/MainUI.prefab
@@ -460,8 +460,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 391.49997, y: -365}
-  m_SizeDelta: {x: 760.16943, y: 0}
+  m_AnchoredPosition: {x: 391.5, y: -365}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &5645569588178486493
 CanvasRenderer:
@@ -1557,7 +1557,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 391.49997, y: -1625}
+  m_AnchoredPosition: {x: 391.5, y: -1625}
   m_SizeDelta: {x: 760.16943, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &3597594261749890995
@@ -2797,7 +2797,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 391.49997, y: -995}
+  m_AnchoredPosition: {x: 391.5, y: -995}
   m_SizeDelta: {x: 760.16943, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &7858627073060222853
@@ -3599,7 +3599,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 391.49997, y: -155}
+  m_AnchoredPosition: {x: 391.5, y: -155}
   m_SizeDelta: {x: 760.16943, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &8739092566329587679
@@ -5761,8 +5761,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 391.49997, y: -1205}
-  m_SizeDelta: {x: 760.16943, y: 0}
+  m_AnchoredPosition: {x: 391.5, y: -1205}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &8577482210663365958
 CanvasRenderer:
@@ -6605,8 +6605,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 391.49997, y: -785}
-  m_SizeDelta: {x: 760.16943, y: 0}
+  m_AnchoredPosition: {x: 391.5, y: -785}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &7695839243741384083
 CanvasRenderer:
@@ -7417,7 +7417,7 @@ MonoBehaviour:
   m_HandleRect: {fileID: 5742302808357039708}
   m_Direction: 2
   m_Value: 1
-  m_Size: 0.2093023
+  m_Size: 0.20922896
   m_NumberOfSteps: 0
   m_OnValueChanged:
     m_PersistentCalls:
@@ -7597,7 +7597,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 391.49997, y: -575}
+  m_AnchoredPosition: {x: 391.5, y: -575}
   m_SizeDelta: {x: 760.16943, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &6178112515859522792
@@ -9257,8 +9257,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 391.49997, y: -1835}
-  m_SizeDelta: {x: 760.16943, y: 0}
+  m_AnchoredPosition: {x: 391.5, y: -1835}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &8999747104183257944
 CanvasRenderer:
@@ -10145,8 +10145,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 391.49997, y: -2045}
-  m_SizeDelta: {x: 760.16943, y: 0}
+  m_AnchoredPosition: {x: 391.5, y: -2045}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &4616489498810665385
 CanvasRenderer:
@@ -11703,8 +11703,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
-  m_AnchoredPosition: {x: 391.49997, y: -1415}
-  m_SizeDelta: {x: 760.16943, y: 0}
+  m_AnchoredPosition: {x: 391.5, y: -1415}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &1731091000113706810
 CanvasRenderer:
