+++ Assets/Scripts/UI/ArrowToWorkbench.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: bc623e8d8a84fea46b524f7245215d42
