--- Assets/Scripts/UI/PointToHotel.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5136f83f8640b9743a661bf25db433a3
